+++ pychron/labspy/client.py
@@ -73,8 +73,7 @@ class LabspyClient(Loggable):
     def add_run(self, run, exp):
         with self.db.session_ctx():
             exp = self.db.get_experiment(self._generate_hid(exp))
-
-            self.db.add_analysis(exp, self._run_dict())
+            self.db.add_analysis(exp, self._run_dict(run))
 
     def add_measurement(self, dev, tag, val, unit):
         with self.db.session_ctx():
@@ -84,28 +83,25 @@ class LabspyClient(Loggable):
     def _get_db(self):
         return LabspyDatabaseAdapter()
 
-    def _run_dict(self, run, exp):
+    def _run_dict(self, run):
         spec = run.spec
 
         d = {dbk: getattr(spec, k) for k, dbk in (('runid', 'Runid'),
-                                                  # ('analysis_type',''),
+                                                  ('analysis_type','analysis_type'),
+                                                  ('analysis_timestamp','TimeStamp'),
                                                   ('sample', 'Sample'),
                                                   ('extract_value', 'ExtractValue'),
                                                   ('extract_units', 'ExtractUnits'),
                                                   ('duration', 'Duration'),
                                                   ('cleanup', 'Cleanup'),
                                                   ('position', 'Position'),
-                                                  # ('comment',),
-                                                  # ('material','Material'),
+                                                  ('comment','Comment'),
+                                                  ('material','Material'),
                                                   ('project', 'Project'),
                                                   ('state', 'State'))}
-        if spec.analysis_timestamp:
-            d['TimeStamp'] = time.mktime(spec.analysis_timestamp.timetuple())
-        else:
-            d['TimeStamp'] = ''
 
         for si in SCRIPT_NAMES:
-            k = ''.join(map(str.capitalize, si.split('_')))
+            k = ''.join(map(str.capitalize, si.split('_')[:-1]))
             d[k] = getattr(spec, si)
 
         return d
@@ -117,236 +113,68 @@ class LabspyClient(Loggable):
         md5.update(exp.starttime.isoformat())
         return md5.hexdigest()
 
+# ================= testing =========================
 
-class MeteorLabspyClient(LabspyClient):
-    host = Str
-    port = Int
-    database_name = Str
-
-    _client = None
-
-    def __init__(self, *args, **kw):
-        super(MeteorLabspyClient, self).__init__(*args, **kw)
-        self.database_name = 'meteor'
-
-    def bind_preferences(self):
-        bind_preference(self, 'host', 'pychron.labspy.host')
-        bind_preference(self, 'port', 'pychron.labspy.port')
-        bind_preference(self, 'database_name', 'pychron.labspy.database_name')
-
-    def test_connection(self, warn=True):
-        if self.host and self.port:
-            from pymongo import MongoClient
-            from pymongo.errors import ConnectionFailure
-
-            url = 'mongodb://{}:{}/'.format(self.host, self.port)
-            try:
-                self._client = MongoClient(url)
-                return True
-            except ConnectionFailure:
-                if warn:
-                    self.warning_dialog('failed connecting to database at {}'.format(url))
-        return False
-
-    def add_experiment(self, exp):
-        db = self.db
-        if db:
-            now = datetime.now()
-            attrs = ('username', 'extract_device', 'mass_spectrometer', 'name', 'status')
-            doc = {ai: getattr(exp, ai) for ai in attrs}
-
-            hid = self._generate_hid(exp)
-            doc.update(**{'starttime': time.mktime(exp.starttime.timetuple()),
-                          'timestamp': time.mktime(now.timetuple()),
-                          'hash_id': hid})
-
-            db.experiments.insert(doc)
-            # exp.hash_id = hid
-
-    def update_experiment(self, exp, err_msg):
-        if self.db:
-            db = self.db
-            hid = self._generate_hid(exp)
-            doc = db.experiments.find_one({'hash_id': hid})
-            if doc:
-                db.experiments.update({'_id': doc['_id']},
-                                      {'$set': {'status': err_msg or 'Finished',
-                                                'timestampf': time.mktime(datetime.now().timetuple())}})
-
-    def update_status(self, **kw):
-        db = self.db
-        if db:
-            doc = db.state.find_one({})
-            if not doc:
-                doc \
-                    = kw
-                db.state.insert(doc)
-            else:
-                db.state.update({'_id': doc['_id']}, {'$set': kw})
-
-    def _generate_hid(self, exp):
-        md5 = hashlib.md5()
-        md5.update(exp.name)
-        md5.update(exp.spectrometer)
-        md5.update(exp.starttime.isoformat())
-        return md5.hexdigest()
-
-    def add_run(self, run, exp):
-        db = self.db
-        if db:
-            # attrs = ('record_id', 'timestamp', 'experiment_name', 'state')
-            # doc = {k: getattr(run, k) for k in attrs}
-            doc = self._make_run_doc(run, exp)
-            db.runs.insert(doc)
-
-    def add_device_post(self, devs):
-        # clt = self._client
-        # if not clt:
-        # return
-        #
-        # db = clt[self.database_name]
-
-        db = self.db
-        if db:
-            now = datetime.now()
-            # print now.isoformat(), now.strftime('%H:%M:%S')
-            doc = {  # 'timestamp': now,
-                     # 'timestampt': now.strftime('%H:%M:%S'),
-                     'timestamp': time.mktime(now.timetuple())}
-
-            values = []
-            for di in devs:
-                for vk, cv, ui in zip(di.value_keys, di.current_values, di.units):
-                    values.append({'device': di.name,
-                                   'name': vk,
-                                   'value': cv,
-                                   'units': ui})
-
-            doc['values'] = values
-
-            # pprint(doc, width=4)
-            db.devices.insert(doc)
-
-    # private
-    def _make_run_doc(self, run, exp):
-        spec = run.spec
-
-        d = {k: getattr(spec, k) for k in ('runid',
-                                           'analysis_type',
-                                           'sample',
-                                           'extract_value',
-                                           'duration',
-                                           'cleanup',
-                                           'position',
-                                           'comment',
-                                           'material',
-                                           'project',
-                                           'mass_spectrometer',
-                                           'extract_device',
-                                           'state')}
-
-        d['experiment_name'] = exp.name
-        if spec.analysis_timestamp:
-            d['timestamp'] = time.mktime(spec.analysis_timestamp.timetuple())
-            # d['date'] = spec.analysis_timestamp.strftime('%m/%d/%Y %I:%M:%S %p')
-            # d['runtime'] = spec.analysis_timestamp.strftime('%I:%M:%S %p')
-        else:
-            d['date'] = ''
-            d['timestamp'] = ''
-            d['runtime'] = ''
-
-        for si in SCRIPT_NAMES:
-            d[si] = getattr(spec, si)
-
-        return d
-
-    def _get_property(self):
-        if self._client:
-            return self._client[self.database_name]
-
-
-def add_device(clt):
-    class Dev():
-        def __init__(self, name, values, units):
-            self.name = name
-            self.units = units
-            self.value_keys = values
-            self.current_values = [random() for k in values]
-
-
-    a = Dev('pneumatic', ('pressure',), ('torr',))
-    b = Dev('environment',
-            ('temperature', 'humidity'), ('C', '%'))
-    c = Dev('gauge', ('bone_ig',), ('torr',))
-    d = Dev('gauge', ('microbone_ig',), ('torr',))
-    clt.add_device_post([a, b, c, d])
-
-
-def add_experiment(c):
-    pass
-    # class Exp():
-    # def __init__(self, name, user, spec, status):
-    # self.name = name
-    # self.username = user
-    # self.spectrometer = spec
-    # self.mass_spectrometer = spec
-    # self.extract_device = choice(('Fusions CO2', 'Fusions Diode'))
-    # self.status = status
-    #         self.starttime = datetime(2014, 11, 1, 12, 10, 10)
-    #
-    # # class Spec():
-    # # def __init__(self, record_id):
-    # #         self.runid = record_id
-    # #         self.mass_spectrometer='jan'
-    # #         self.extract_device='LF'
-    # #         self.analysis_timestamp = datetime.now()
-    # #         self.state = choice(['Finished', 'Canceled', 'Failed'])
-    # #         self.analysis_type = "unknown"
-    # #         self.sample = "FC-2"
-    # #         self.extract_value = random()*2
-    # #         self.duration = randint(100,200)
-    # #         self.cleanup = randint(100,200)
-    # #         self.position = 1
-    # #         self.comment = "Test comment"
-    # #         self.material = "sanidine"
-    # #         self.project = "Monitor"
-    # #         self.measurement_script = 'm'
-    # #         self.extraction_script = 'e'
-    # #         self.post_measurement_script = 'pm'
-    # #         self.post_equilibration_script = 'pq'
-    # #
-    # # class Run():
-    # #     def __init__(self, *args, **kw):
-    # #         self.spec = Spec(*args, **kw)
-    # #
-    # #
-    # e = Exp('Current Experiment', 'foobar', 'Jan', 'Running')
-    # c.add_experiment(e)
-    #
-    # # # print e.hash_id
-    # # # hid='076441e14fe0e09086626f25f216ca04'
-    # # # e.hash_id=hid
-    # # # c.update_experiment(e)
-    # #
-    # # for i in range(6):
-    # #     c.add_run(Run('20016-{:02n}'.format(i + 1)), e)
-
-
-def update_status(c):
-    c.update_status(error='Error big time')
-
-
-def add_measurements(c):
-    for i in range(10):
-        c.add_measurement('AirPressure', 'pneumatics', random(), 'PSI')
-        c.add_measurement('Environmental', 'temperature', random() * 100, 'C')
-        c.add_measurement('Environmental', 'humidity', random() * 100, '%')
-
-        time.sleep(0.25)
 
 
 if __name__ == '__main__':
     from random import random, choice, randint
+
+    def add_runs(c, e):
+        class Spec():
+            def __init__(self, record_id):
+                self.runid = record_id
+                self.mass_spectrometer='jan'
+                self.extract_device='LF'
+                self.analysis_timestamp = datetime.now()
+                self.state = choice(['Finished', 'Canceled', 'Failed'])
+                self.analysis_type = "unknown"
+                self.sample = "FC-2"
+                self.extract_value = random()*2
+                self.extract_units = 'watts'
+                self.duration = randint(100,200)
+                self.cleanup = randint(100,200)
+                self.position = 1
+                self.comment = "Test comment"
+                self.material = "sanidine"
+                self.project = "Monitor"
+                self.measurement_script = 'm'
+                self.extraction_script = 'e'
+                self.post_measurement_script = 'pm'
+                self.post_equilibration_script = 'pq'
+
+        class Run():
+            def __init__(self, *args, **kw):
+                self.spec = Spec(*args, **kw)
+
+        for i in range(6):
+            c.add_run(Run('20016-{:02n}'.format(i + 1)), e)
+
+    def add_experiment(c):
+        class Exp():
+            def __init__(self, name, user, spec, status):
+                self.name = name
+                self.username = user
+                self.spectrometer = spec
+                self.mass_spectrometer = spec
+                self.extract_device = choice(('Fusions CO2', 'Fusions Diode'))
+                self.status = status
+                self.starttime = datetime.now()
+        e = Exp('Current Experiment', 'foobar', 'Jan', 'Running')
+        c.add_experiment(e)
+        return e
+
+    def update_status(c):
+        c.update_status(Error='Error big time')
+        c.update_status(Message='This is a long message', ShortMessage='This is a short message')
+
+    def add_measurements(c):
+        for i in range(10):
+            c.add_measurement('AirPressure', 'pneumatics', random(), 'PSI')
+            c.add_measurement('Environmental', 'temperature', random() * 100, 'C')
+            c.add_measurement('Environmental', 'humidity', random() * 100, '%')
+
+            time.sleep(0.25)
 
     logging_setup('labspyclient')
     # c = LabspyClient(bind=False, host='129.138.12.138', port=27017)
@@ -356,15 +184,185 @@ if __name__ == '__main__':
     # add_device(c)
     # time.sleep(1)
 
-    c = LabspyClient(bind=False)
-    c.db.host = 'localhost'
-    c.db.username = 'root'
-    c.db.password = 'Argon'
-    c.db.name = 'argonlab'
-    c.test_connection()
-    add_measurements(c)
+    clt = LabspyClient(bind=False)
+    clt.db.host = 'localhost'
+    clt.db.username = 'root'
+    clt.db.password = 'Argon'
+    clt.db.name = 'argonlab'
+    clt.test_connection()
 
+    # set status
+    update_status(clt)
+
+    # measurements
+    # add_measurements(clt)
+
+    # # experiments/runs
+    # exp = add_experiment(clt)
+    # add_runs(clt, exp)
 # ============= EOF =============================================
+# class MeteorLabspyClient(LabspyClient):
+#     host = Str
+#     port = Int
+#     database_name = Str
+#
+#     _client = None
+#
+#     def __init__(self, *args, **kw):
+#         super(MeteorLabspyClient, self).__init__(*args, **kw)
+#         self.database_name = 'meteor'
+#
+#     def bind_preferences(self):
+#         bind_preference(self, 'host', 'pychron.labspy.host')
+#         bind_preference(self, 'port', 'pychron.labspy.port')
+#         bind_preference(self, 'database_name', 'pychron.labspy.database_name')
+#
+#     def test_connection(self, warn=True):
+#         if self.host and self.port:
+#             from pymongo import MongoClient
+#             from pymongo.errors import ConnectionFailure
+#
+#             url = 'mongodb://{}:{}/'.format(self.host, self.port)
+#             try:
+#                 self._client = MongoClient(url)
+#                 return True
+#             except ConnectionFailure:
+#                 if warn:
+#                     self.warning_dialog('failed connecting to database at {}'.format(url))
+#         return False
+#
+#     def add_experiment(self, exp):
+#         db = self.db
+#         if db:
+#             now = datetime.now()
+#             attrs = ('username', 'extract_device', 'mass_spectrometer', 'name', 'status')
+#             doc = {ai: getattr(exp, ai) for ai in attrs}
+#
+#             hid = self._generate_hid(exp)
+#             doc.update(**{'starttime': time.mktime(exp.starttime.timetuple()),
+#                           'timestamp': time.mktime(now.timetuple()),
+#                           'hash_id': hid})
+#
+#             db.experiments.insert(doc)
+#             # exp.hash_id = hid
+#
+#     def update_experiment(self, exp, err_msg):
+#         if self.db:
+#             db = self.db
+#             hid = self._generate_hid(exp)
+#             doc = db.experiments.find_one({'hash_id': hid})
+#             if doc:
+#                 db.experiments.update({'_id': doc['_id']},
+#                                       {'$set': {'status': err_msg or 'Finished',
+#                                                 'timestampf': time.mktime(datetime.now().timetuple())}})
+#
+#     def update_status(self, **kw):
+#         db = self.db
+#         if db:
+#             doc = db.state.find_one({})
+#             if not doc:
+#                 doc \
+#                     = kw
+#                 db.state.insert(doc)
+#             else:
+#                 db.state.update({'_id': doc['_id']}, {'$set': kw})
+#
+#     def _generate_hid(self, exp):
+#         md5 = hashlib.md5()
+#         md5.update(exp.name)
+#         md5.update(exp.spectrometer)
+#         md5.update(exp.starttime.isoformat())
+#         return md5.hexdigest()
+#
+#     def add_run(self, run, exp):
+#         db = self.db
+#         if db:
+#             # attrs = ('record_id', 'timestamp', 'experiment_name', 'state')
+#             # doc = {k: getattr(run, k) for k in attrs}
+#             doc = self._make_run_doc(run, exp)
+#             db.runs.insert(doc)
+#
+#     def add_device_post(self, devs):
+#         # clt = self._client
+#         # if not clt:
+#         # return
+#         #
+#         # db = clt[self.database_name]
+#
+#         db = self.db
+#         if db:
+#             now = datetime.now()
+#             # print now.isoformat(), now.strftime('%H:%M:%S')
+#             doc = {  # 'timestamp': now,
+#                      # 'timestampt': now.strftime('%H:%M:%S'),
+#                      'timestamp': time.mktime(now.timetuple())}
+#
+#             values = []
+#             for di in devs:
+#                 for vk, cv, ui in zip(di.value_keys, di.current_values, di.units):
+#                     values.append({'device': di.name,
+#                                    'name': vk,
+#                                    'value': cv,
+#                                    'units': ui})
+#
+#             doc['values'] = values
+#
+#             # pprint(doc, width=4)
+#             db.devices.insert(doc)
+#
+#     # private
+#     def _make_run_doc(self, run, exp):
+#         spec = run.spec
+#
+#         d = {k: getattr(spec, k) for k in ('runid',
+#                                            'analysis_type',
+#                                            'sample',
+#                                            'extract_value',
+#                                            'duration',
+#                                            'cleanup',
+#                                            'position',
+#                                            'comment',
+#                                            'material',
+#                                            'project',
+#                                            'mass_spectrometer',
+#                                            'extract_device',
+#                                            'state')}
+#
+#         d['experiment_name'] = exp.name
+#         if spec.analysis_timestamp:
+#             d['timestamp'] = time.mktime(spec.analysis_timestamp.timetuple())
+#             # d['date'] = spec.analysis_timestamp.strftime('%m/%d/%Y %I:%M:%S %p')
+#             # d['runtime'] = spec.analysis_timestamp.strftime('%I:%M:%S %p')
+#         else:
+#             d['date'] = ''
+#             d['timestamp'] = ''
+#             d['runtime'] = ''
+#
+#         for si in SCRIPT_NAMES:
+#             d[si] = getattr(spec, si)
+#
+#         return d
+#
+#     def _get_property(self):
+#         if self._client:
+#             return self._client[self.database_name]
+#
+#
+# def add_device(clt):
+#     class Dev():
+#         def __init__(self, name, values, units):
+#             self.name = name
+#             self.units = units
+#             self.value_keys = values
+#             self.current_values = [random() for k in values]
+#
+#
+#     a = Dev('pneumatic', ('pressure',), ('torr',))
+#     b = Dev('environment',
+#             ('temperature', 'humidity'), ('C', '%'))
+#     c = Dev('gauge', ('bone_ig',), ('torr',))
+#     d = Dev('gauge', ('microbone_ig',), ('torr',))
+#     clt.add_device_post([a, b, c, d])